Fix selections in Atom
They made a change that would include dashes as part of a word. So if
you had something like `foo-bar-baz` and you double clicked on "baz" the
entire string would get highlighted. This is not ideal when you're
working with CSS classnames as you generally only want to modify the one
part of the string that you're double clicking.

The change here takes the existing list of non word characters and adds
the dash back in.